Update user cart

+++ backend/src/main/java/com/furnifinders/backend/dto/Request/PostProductRequest.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class PostProductRequest {
     private String product_name;
     private String product_description;
-    private int product_price;
-    private int product_percentage;
-    private int product_quantity;
+    private Long product_price;
+    private Long product_percentage;
+    private Long product_quantity;
 }
